Move sphere related code into it's own file
